+++ src/theme/debug.js
@@ -2,6 +2,8 @@ const { lightness } = require('../base/util')
 
 exports.debug = function (color) {
     return {
-        'debugToolBar.background': lightness(color.bg, -2)
+        'debugToolBar.background': lightness(color.bg, -2),
+
+        'debugIcon.breakpointForeground': color.red
     }
 }